+++ platform/darwin/src/MGLRendererConfiguration.mm
@@ -1,5 +1,6 @@
 #import "MGLRendererConfiguration.h"
 #import "MGLOfflineStorage_Private.h"
+#import "MGLFoundation_Private.h"
 
 #if TARGET_OS_IPHONE
 #import <UIKit/UIKit.h>
@@ -7,11 +8,53 @@
 #import <AppKit/AppKit.h>
 #endif
 
+static NSString * const MGLCollisionBehaviorPre4_0Key = @"MGLCollisionBehaviorPre4_0";
+
+@interface MGLRendererConfiguration ()
+@property (nonatomic, readwrite) BOOL perSourceCollisions;
+@end
+
 
 @implementation MGLRendererConfiguration
 
 + (instancetype)currentConfiguration {
     return [[self alloc] init];
+}
+
+- (instancetype)init {
+    return [self initWithPropertyDictionary:[[NSBundle mainBundle] infoDictionary]];
+}
+
+- (instancetype)initWithPropertyDictionary:(nonnull NSDictionary *)properties {
+    self = [super init];
+    
+    if (self) {
+        // Set the collision behaviour. A value set in `NSUserDefaults.standardUserDefaults`
+        // should override anything in the application's info.plist
+        NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
+
+        if ([defaults objectForKey:MGLCollisionBehaviorPre4_0Key]) {
+            _perSourceCollisions = [defaults boolForKey:MGLCollisionBehaviorPre4_0Key];
+        }
+        else {
+            id collisionBehaviourValue = properties[MGLCollisionBehaviorPre4_0Key];
+            
+            NSNumber *collisionBehaviourNumber = MGL_OBJC_DYNAMIC_CAST(collisionBehaviourValue, NSNumber);
+            
+            if (collisionBehaviourNumber) {
+                _perSourceCollisions = collisionBehaviourNumber.boolValue;
+            } else {
+                // Also support NSString to correspond with the behavior of `-[NSUserDefaults boolForKey:]`
+                NSString *collisionBehaviourString = MGL_OBJC_DYNAMIC_CAST(collisionBehaviourValue, NSString);
+
+                if (collisionBehaviourString) {
+                    _perSourceCollisions = collisionBehaviourString.boolValue;
+                }
+            }
+        }
+    }
+    
+    return self;
 }
 
 - (mbgl::DefaultFileSource *)fileSource {
@@ -38,12 +81,6 @@
     NSString *fontFamilyName = [[NSBundle mainBundle] objectForInfoDictionaryKey:@"MGLIdeographicFontFamilyName"];
 
     return fontFamilyName ? std::string([fontFamilyName UTF8String]) : mbgl::optional<std::string>();
-}
-
-- (BOOL)perSourceCollisions {
-    NSNumber *boolWrapper = [[[NSBundle mainBundle] infoDictionary] valueForKey:@"MGLCollisionBehaviorPre4_0"];
-    
-    return boolWrapper.boolValue;
 }
 
 @end